added descriptions to some functions
